Fixed warning // dangling pointer

+++ ReVengine/ReVengine-Engine/Rendering/Shaders/BaseShader.cpp
@@ -16,7 +16,7 @@ void Rev::BaseShader::SetupInputLayer()
 		{ "POSITION", 0, DXGI_FORMAT_R32G32B32_FLOAT, 0, D3D11_APPEND_ALIGNED_ELEMENT, D3D11_INPUT_PER_VERTEX_DATA, 0 },
 		{ "UV", 0, DXGI_FORMAT_R32G32_FLOAT, 0, D3D11_APPEND_ALIGNED_ELEMENT, D3D11_INPUT_PER_VERTEX_DATA, 0 },
 	};
-	m_Device->CreateInputLayout(inputElement_DESC, std::size(inputElement_DESC), m_VertexBytecode.c_str(), m_VertexBytecode.size(), &m_InputLayer);
+	m_Device->CreateInputLayout(inputElement_DESC, UINT(std::size(inputElement_DESC)), m_VertexBytecode.c_str(), m_VertexBytecode.size(), &m_InputLayer);
 }
 
 void BaseShader::SetShaderStages()
